cleanup

+++ problems/Problem_210.hpp
@@ -10,7 +10,7 @@ It is conjectured that every such sequence eventually reaches the number 1. Test
 Bonus: What input n <= 1000000 gives the longest sequence?
 */
 
-inline long int collatz( int n )
+inline long int collatz( long int n )
 {
   long int count = 0;
   while( n != 1 )
